// frontend/magic-web/src/opensource/pages/superMagic/pages/Workspace/types.ts
export enum TopicMode {
	/** 通用模式 */
	General = "general",
	/** 聊天模式 */
	Chat = "chat",
	/** 数据分析 */
	DataAnalysis = "data_analysis",
	/** PPT */
	PPT = "ppt",
	/** 研报模式 */
	Report = "report",
	/** 录音总结 */
	RecordSummary = "summary",
	/** 空模式 */
	Empty = "",
	/** 设计模式 */
	Design = "design",
	/** 员工主导模式
	 * 不会在员工列表中存在，只用于创建员工时使用
	 */
	AgentManager = "agent-manager",
}